Move explorer module inside the api
