+++ exonum/src/lib.rs
@@ -82,5 +82,4 @@ pub mod node;
 pub mod storage;
 #[macro_use]
 pub mod blockchain;
-pub mod explorer;
 pub mod api;
