+++ testkit/examples/timestamping.rs
@@ -20,7 +20,7 @@ extern crate serde_json;
 
 use exonum::crypto::{gen_keypair, Hash, PublicKey, CryptoHash};
 use exonum::blockchain::{Schema, Service, Transaction, TransactionSet, ExecutionResult};
-use exonum::explorer::BlocksRange;
+use exonum::api::public::BlocksRange;
 use exonum::messages::{Message, RawTransaction};
 use exonum::storage::{Fork, Snapshot};
 use exonum::encoding;
